Remove RunAsync annotation Haulmont/jmix-ui#156
Remove RunAsync annotation and add app property for timeout for all operations.
Remove - JmxConsoleAutoConfiguration, AnnotationMBeanInfoAssembler,
JmxConsoleMBeanExporter.
Add a progress bar on MBeanOperationResultScreen that is shown until
invoked operation is finished or timeout is exceeded.
After task is finished, result is shown on MBeanOperationResultScreen.
If timeout is exceeded, MBeanOperationResultScreen is closed and
notification is shown.