Fix columns order in excelExport action Haulmont/jmix-ui#784
Replacing call of `table.getColumns` with `table.getNotCollapsedColumns()` because for export functionality it is necessary to use _visible columns_ with appropriate order instead of columns with order as they were parsed.
Note, that `table.getColumns` will return list of columns with next ordering - columns defined as '<column>' elements, after that columns that are defined inside '<group>' element. It is not always correlated with order of visible columns (see io.jmix.ui.xml.layout.loader.GroupTableLoader.loadColumns for details)