A method has been added to the Downloader interface that accepts a delegate, which is a function that accepts the file extension as a String and returns a boolean verdict: whether the value is included in a variety of allowed formats or not (whether to display this document inline or not)
If the delegate is not inserted, the default behavior set works for UiProperties.

In the DownloaderImpl class, methods have been added that provide default behavior in case the delegate has not been passed.

In the UiProperties class, the htm and html file extensions have been removed from the default lists.

A list of default extensions for the Reports addon has been added to the Report Properties class. As before, it contains html and html default formats.

The ReportProperties dependency has been injected in the View classes ReportListView and ReportsDetailView, based on which a delegate is set in the Downloader bean. The delegate is set in the OnInit methods of these classes